# api-spec.yaml
openapi: 3.0.3
info:
  title: POS Stripe Connect API
  description: |
    API for managing Stripe Connect integration for POS systems.

    This API provides endpoints for:
    - User authentication and authorization
    - Store management
    - Customer management
    - POS device registration and management
    - POS session management (Kassasystemforskriften compliance)
    - POS event logging (audit trail)
    - POS transaction operations (void, correction)
    - Receipt generation and management
    - Receipt printer configuration and management
    - Product and inventory management
    - SAF-T file generation (Norwegian tax compliance)
    - Terminal operations (connection tokens and payment intents)

    All endpoints (except login and webhooks) require Bearer token authentication.
    Requests are automatically scoped to the authenticated user's accessible stores.
  version: 1.0.0
  contact:
    name: API Support
    email: support@visivo.no
  license:
    name: Proprietary
servers:
  - url: https://pos.visivo.no/api
    description: Production server
  - url: http://pos-stripe.test/api
    description: Local development server
...
paths:
...
  /products/{product}/images/{media}:
    get:
      tags:
        - Products
      summary: Serve product image
      description: Serve product image with signed URL (public endpoint, secured by signature)
      operationId: serveProductImage
      security: []
      parameters:
        - name: product
          in: path
          required: true
          schema:
            type: integer
        - name: media
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product image
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: Product or image not found
        '403':
          description: Invalid signature
...
  /collections:
    get:
      tags:
        - Collections
      summary: List collections
      description: 'Get paginated list of collections. If there are products with no collection, a fake "Ukategorisert" category with id 0 will be included in the list.'
      operationId: listCollections
      parameters:
        - name: search
          in: query
          description: Search term (name or description)
          required: false
          schema:
            type: string
        - name: active
          in: query
          description: Filter by active status (defaults to true)
          required: false
          schema:
            type: boolean
        - name: per_page
          in: query
          description: Number of items per page (max 100)
          required: false
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: List of collections
          content:
            application/json:
              schema:
                type: object
                properties:
                  collections:
                    type: array
                    items:
                      $ref: '#/components/schemas/Collection'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '404':
          description: Store not found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /collections/{id}:
    get:
      tags:
        - Collections
      summary: Get collection
      description: Get a specific collection with its products
      operationId: getCollection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Collection information with products
          content:
            application/json:
              schema:
                type: object
                properties:
                  collection:
                    allOf:
                      - $ref: '#/components/schemas/Collection'
                      - type: object
                        properties:
                          products:
                            type: array
                            items:
                              $ref: '#/components/schemas/Product'
        '404':
          description: Collection not found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Sanctum Bearer Token authentication.

        Include the token in the Authorization header:
        ```
        Authorization: Bearer {token}
        ```

        Get a token by calling the `/auth/login` endpoint.
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        current_page:
          type: integer
          example: 1
        last_page:
          type: integer
          example: 10
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 200
...
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        stripe_product_id:
          type: string
          nullable: true
          example: prod_123
        name:
          type: string
          example: Coffee
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [service, good]
          example: good
        active:
          type: boolean
          example: true
        shippable:
          type: boolean
          example: false
        url:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
            format: uri
          description: Array of image URLs (signed URLs for security)
        no_price_in_pos:
          type: boolean
          description: If true, this product has no preset price and requires custom price input on POS.
          default: false
        product_price:
          type: object
          nullable: true
          description: Default/active price
          properties:
            id:
              type: string
            amount:
              type: integer
              description: Price in øre
            amount_formatted:
              type: string
            currency:
              type: string
            type:
              type: string
            is_default:
              type: boolean
        prices:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              amount:
                type: integer
              amount_formatted:
                type: string
              currency:
                type: string
              type:
                type: string
              is_default:
                type: boolean
        variants:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              stripe_product_id:
                type: string
                nullable: true
              stripe_price_id:
                type: string
                nullable: true
              sku:
                type: string
                nullable: true
              barcode:
                type: string
                nullable: true
              variant_name:
                type: string
              variant_options:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    value:
                      type: string
              variant_price:
                type: object
                description: Variant price object (for FlutterFlow compatibility)
                properties:
                  amount:
                    type: integer
                    description: Price in øre. Returns 0 for variants without preset prices (custom price input on POS).
                  amount_formatted:
                    type: string
                    description: Formatted price string. Returns "0.00" for variants without preset prices.
              price_amount:
                type: integer
                nullable: false
                description: Price in øre (flattened field for backward compatibility). Returns 0 for variants without preset prices (custom price input on POS). Frontend should check if price_amount === 0 to enable custom price input.
              price_amount_formatted:
                type: string
                nullable: false
                description: Formatted price string (flattened field for backward compatibility). Returns "0.00" for variants without preset prices. Frontend should check if price_amount === 0 to enable custom price input.
              compare_at_price_amount:
                type: integer
                nullable: true
                description: Compare at price in øre (original price for discounts).
              compare_at_price_amount_formatted:
                type: string
                nullable: true
                description: Formatted compare at price string.
              currency:
                type: string
              no_price_in_pos:
                type: boolean
                description: If true, this variant has no preset price and requires custom price input on POS.
                default: false
              variant_inventory:
                type: object
                properties:
                  quantity:
                    type: integer
                    nullable: true
                  in_stock:
                    type: boolean
                  policy:
                    type: string
                    nullable: true
                  management:
                    type: string
                    nullable: true
                  tracked:
                    type: boolean
        variants_count:
          type: integer
        product_inventory:
          type: object
          properties:
            tracked:
              type: boolean
            total_quantity:
              type: integer
              nullable: true
            in_stock_variants:
              type: integer
            out_of_stock_variants:
              type: integer
            all_in_stock:
              type: boolean
              nullable: true
        tax_code:
          type: string
          nullable: true
        unit_label:
          type: string
          nullable: true
        statement_descriptor:
          type: string
          nullable: true
        collections:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              handle:
                type: string
                nullable: true
          description: Collections this product belongs to
        package_dimensions:
          type: object
          nullable: true
        product_meta:
          type: object
          nullable: true
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Collection:
      type: object
      description: 'Collection object. Note: id 0 represents a fake "Ukategorisert" (Uncategorized) category that appears when there are products with no collection.'
      properties:
        id:
          type: integer
          description: 'Collection ID. Value 0 indicates the fake "Ukategorisert" category for uncategorized products.'
          example: 1
        name:
          type: string
          example: Summer Collection
        description:
          type: string
          nullable: true
          example: Products for summer season
        handle:
          type: string
          nullable: true
          example: summer-collection
        image_url:
          type: string
          nullable: true
          format: uri
        active:
          type: boolean
          example: true
        sort_order:
          type: integer
          example: 0
        products_count:
          type: integer
          description: Number of products in this collection
          example: 15
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Error message
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email:
              - The email field is required.
...
  responses:
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                message: Unauthenticated.
security:
  - bearerAuth: []